Updated verify to accept CA file and fixed tests.
Verify now requires a CA file to validate certificates. This feature
broke the existing verify test so the tests were rewritten. The
tests have been rewritten to only exercise the functionality of the
library rather than using a bunch of fixtures to exercise the
integration of the SSL module and the OpenSSL command line utility.
The test framework has been changed to use Jasmine rather than
Mocha to take advantage of the spy feature; the make file has been
updated accordingly. This upgrades the version to 1.0.0. The README
has been updated to remove some errors and reflect the changes in
this version.